+++ src/Evaluator/Evaluator.cpp
@@ -329,11 +329,11 @@ ObjPointer Evaluator::evaluate(ASTPointer ast) {
     // builtin func
     if (cb->builtin) {
 
-      if (args.size() < cb->builtin->argcount) {
+      if ((int)args.size() < cb->builtin->argcount) {
         Error(ast->token, "too few arguments")();
       }
       else if (!cb->builtin->is_variable_args &&
-               args.size() > cb->builtin->argcount) {
+               (int)args.size() > cb->builtin->argcount) {
         Error(ast->token, "too many arguments")();
       }
